Split off FontRulesetIcons
